+++ pyodide_build/xbuildenv_releases.py
@@ -15,6 +15,9 @@ NIGHTLY_CROSS_BUILD_ENV_METADATA_URL = (
 NIGHTLY_DEBUG_CROSS_BUILD_ENV_METADATA_URL = (
     "https://pyodide.github.io/pyodide-build-environment-nightly/api/v2/debug.json"
 )
+STABLE_DEBUG_CROSS_BUILD_ENV_METADATA_URL = (
+    "https://pyodide.github.io/pyodide/api/v2/debug.json"
+)
 CROSS_BUILD_ENV_METADATA_URL_ENV_VAR = "PYODIDE_CROSS_BUILD_ENV_METADATA_URL"